third_party/blink/web_tests/external/wpt/css/css-shapes/shape-outside/values/shape-outside-circle-010-expected.txt

This is a testharness.js-based test.
FAIL circle(calc(10in)) - inline style assert_equals: expected "circle(calc(960px) at 50% 50%)" but got "circle(calc(10in))"
FAIL circle(calc(10in + 20px)) - inline style assert_equals: expected "circle(calc(980px) at 50% 50%)" but got "circle(calc(980px))"
FAIL circle(calc(30%)) - inline style assert_equals: expected "circle(calc(30%) at 50% 50%)" but got "circle(calc(30%))"
FAIL circle(calc(100%/4)) - inline style assert_equals: expected "circle(calc(25%) at 50% 50%)" but got "circle(calc(25%))"
FAIL circle(calc(25%*3)) - inline style assert_equals: expected "circle(calc(75%) at 50% 50%)" but got "circle(calc(75%))"
FAIL circle(calc(25%*3 - 10in)) - inline style assert_equals: expected "circle(calc(75% - 960px) at 50% 50%)" but got "circle(calc(75% - 10in))"
FAIL circle(calc((12.5%*6 + 10in) / 4)) - inline style assert_equals: expected "circle(calc(18.75% + 240px) at 50% 50%)" but got "circle(calc((75% + 10in) / 4))"
FAIL circle(calc(10in)) - computed style assert_equals: expected "circle(960px at 50% 50%)" but got "circle(960px)"
FAIL circle(calc(10in + 20px)) - computed style assert_equals: expected "circle(980px at 50% 50%)" but got "circle(980px)"
FAIL circle(calc(30%)) - computed style assert_equals: expected "circle(30% at 50% 50%)" but got "circle(30%)"
FAIL circle(calc(100%/4)) - computed style assert_equals: expected "circle(25% at 50% 50%)" but got "circle(25%)"
FAIL circle(calc(25%*3)) - computed style assert_equals: expected "circle(75% at 50% 50%)" but got "circle(75%)"
FAIL circle(calc(25%*3 - 10in)) - computed style assert_equals: expected "circle(calc(75% - 960px) at 50% 50%)" but got "circle(calc(75% - 960px))"
FAIL circle(calc((12.5%*6 + 10in) / 4)) - computed style assert_equals: expected "circle(calc(18.75% + 240px) at 50% 50%)" but got "circle(calc(18.75% + 240px))"
Harness: the test ran to completion.
